+++ routes/gallery.js
@@ -1,32 +1,24 @@
 //Express
-const express = require('express')
-
-//Base Functions
-const functions = require('./function')
+const express = require('express');
 
 //Authentication 
-const checkAuthenticated = functions.checkAuthenticated
-
-//Unix Time Conversion
-const timeConverter = functions.timeConverter
+const { checkAuthenticated, timeConverter } = require('./engines/functions/function');
 
 //Mysql Functionality
-const mysql = require('mysql2')
+const mysql = require('mysql2');
 
 //Database Configuration
-const database = require('./database.js')
-const db = database.db
+const { db } = require('./database.js');
 
-//Multer for Disk Storage Pathing
-const multer  = require('multer')
+//Uploading Engine for S3
+const { upload }  = require('./engines/gallery/galleryImages');
 
-    //Uploading Engine
-    const uploading = require('./engines/galleryImages')
+//DeleteKey Function 
+const { deleteKey } = require('./engines/gallery/galleryDelete');
 
-    const galleryUpload = uploading.galleryUpload
-
-//FS for File Directory Pathing, may be Unnecessary
-const fs = require('fs') 
+//Get the UserID 
+require('dotenv').config();
+const userID = process.env.ID;
 
 module.exports = app => {
 
@@ -34,48 +26,48 @@ module.exports = app => {
 
     //Public Gallery
     app.get('/gallery', (req, res) => {
-    db.getConnection( async (err, connection) => {
-      if (err) throw (err)
-      const sqlSearch = "SELECT * FROM gallery;"
-      const galleryResult = mysql.format(sqlSearch)
-  
-      connection.query (galleryResult, (err, result) => {
-        connection.release()
-        if (err) throw (err)
-  
-        const gallery = result
-  
-        res.render('gallery',
-        {
-           gallery, 
+      db.getConnection( async (err, connection) => {
+        if (err) throw (err);
+        const sqlSearch = "SELECT * FROM gallery WHERE userID = ?;";
+        const galleryResult = mysql.format(sqlSearch, userID);
 
-          //Header
-          originalUrl : req.originalUrl,
-          isAuthenticated : req.isAuthenticated(),
+        connection.query (galleryResult, (err, result) => {
+          connection.release();
+          if (err) throw (err);
+    
+          const gallery = result;
 
-          timeConverter 
+          res.render('gallery',
+          {
+            gallery : gallery, 
 
+            //Header
+            originalUrl : req.originalUrl,
+            isAuthenticated : req.isAuthenticated(),
+
+            timeConverter 
+
+          })
         })
       })
     })
-  })
   
     //Photo Editing Page
   app.get('/editGallery', checkAuthenticated, (req, res) => {
     db.getConnection( async (err, connection) => {
-      if (err) throw (err)
-      const sqlSearch = "SELECT * FROM gallery;"
-      const galleryResult = mysql.format(sqlSearch)
+      if (err) throw (err);
+      const sqlSearch = "SELECT * FROM gallery WHERE userID = ?;";
+      const galleryResult = mysql.format(sqlSearch, userID);
   
       connection.query (galleryResult, (err, result) => {
-        connection.release()
-        if (err) throw (err)
+        connection.release();
+        if (err) throw (err);
   
-        const gallery = result
+        const gallery = result;
   
         res.render('editGallery', 
         { 
-          gallery, 
+          gallery : gallery, 
 
           //Header
           originalUrl : req.originalUrl,
@@ -90,44 +82,44 @@ module.exports = app => {
   //*********POST REQUESTS*********//
   
   app.post('/photoUpload', 
-   galleryUpload.array('galleryImages'),
+   upload.array('galleryImages'),
    checkAuthenticated, 
    async (req, res) => {
     try {
   
-      const date = req.body.newDate
+      const date = req.body.newDate;
   
-      const fileAmount = req.files.length 
-  
+      const fileAmount = req.files.length;
+
       await new Promise ((resolve,reject) => {
         db.getConnection( async (err, connection) => {
-        if (err) throw (err)
+        if (err) throw (err);
   
-        const sqlInsert = "INSERT INTO gallery (uploadedAt, name) VALUES ?"
+        const sqlInsert = "INSERT INTO gallery (uploadedAt, name, userID) VALUES ?";
   
-        const values = []
+        const values = [];
         
         for (let i = 0; i < fileAmount; i++) {
-          values.push([date, req.files[i].filename])
+          values.push([date, req.files[i].originalname, userID]);
         }
 
-        const insertImages = mysql.format(sqlInsert, [values])
+        const insertImages = mysql.format(sqlInsert, [values]);
   
         connection.query (insertImages, (err, result) => {
-          connection.release()
-          if (err) reject(err)
-          console.log(err,result)
-          return resolve(result)
+          connection.release();
+          if (err) reject(err);
+          console.log(err,result);
+          return resolve(result);
           })
         })
       })
   
-    res.redirect('/editGallery')
+    res.redirect('/editGallery');
   
     } catch {
   
-    res.redirect('/editGallery')
-    console.log('ERROR')
+    res.redirect('/editGallery');
+    console.log('ERROR');
   
     } 
   })
@@ -135,24 +127,24 @@ module.exports = app => {
   app.put('/photoUpdate', checkAuthenticated, async (req, res) => {
     try {
 
-      console.log(req.body)
-      const response = req.body
+      console.log(req.body);
+      const response = req.body;
 
       const valuesArray = [];
 
       for (const key in response) {
-        const values = response[key]
+        const values = response[key];
 
-        valuesArray.push(values)
+        valuesArray.push(values);
 
       }
 
       await new Promise ((resolve,reject) => {
         db.getConnection( async (err,connection) => {
 
-          if (err) reject(err)
+          if (err) reject(err);
 
-          count = 0
+          count = 0;
 
           //Experimenting with Possible Insertion Techniques
           //loop through the array in steps of 3
@@ -161,31 +153,31 @@ module.exports = app => {
 
           for (let i = 0; i < valuesArray.length; i += 3) {
 
-            const caption = valuesArray[i]
-            const date = valuesArray[i + 1]
-            const name = valuesArray[i + 2]
+            const caption = valuesArray[i];
+            const date = valuesArray[i + 1];
+            const name = valuesArray[i + 2];
 
-            console.log(caption, date, name)
+            console.log(caption, date, name);
 
             const capUpdate = `UPDATE gallery
             SET 
               caption = ?
-            WHERE uploadedAt = ? AND name = ?`
+            WHERE uploadedAt = ? AND name = ? AND userID = ?`;
 
-            const capUpdateQuery = mysql.format(capUpdate, [caption, date, name])
+            const capUpdateQuery = mysql.format(capUpdate, [caption, date, name, userID]);
             
             // execute the MySQL query for this set of values
             connection.query(capUpdateQuery, (err, result) => {
 
-              connection.release()
+              connection.release();
 
               if (err) reject(err)
-              console.log(`Updated caption for ${name}`)
+              console.log(`Updated caption for ${name}`);
 
               count++
             
               if (count === valuesArray.length/3) {
-                return resolve(result)
+                return resolve(result);
               }
 
             });
@@ -196,30 +188,32 @@ module.exports = app => {
 
       })
 
-      res.redirect('/gallery')
+      res.redirect('/gallery');
 
     } catch {
 
-      res.redirect('/editGallery')
+      res.redirect('/editGallery');
 
     }
   })
-  
+
   app.delete('/photoDelete', checkAuthenticated, async (req, res) => {
     try {
 
       //Condense req.body for readability 
-      const response = req.body
+      const response = req.body;
+
+      console.log(response);
   
       //Get Keys to Match ID Column in MySQL DB
-      const keysArray = []
+      const keysArray = [];
   
       //Keys Array Needs to Equal Names Array for Loop and File Delete to Work Together
-      const namesArray = []
+      const namesArray = [];
       
       for (const key in response) {
         const values = response[key];
-        const valuesArray = []
+        const valuesArray = [];
         //Obj Values are Strings if 1:1 with Object Keys, e.g. { key : name }
         //Obj keys need to be 1:1 due to the way the way the files are deleted, i.e.,
         //the function looks for the folder (UNIX timestamp value held in the DB) AND the file name
@@ -241,45 +235,46 @@ module.exports = app => {
         await new Promise ((resolve,reject) => {
 
           db.getConnection( async (err, connection) => {
-            if (err) reject(err)
+            if (err) reject(err);
   
-            //Delete By Photo ID, Unique Auto-Incrementing Value
-            const photoDelete = 'DELETE FROM gallery WHERE uploadedAt = ? AND name = ?'
+            //Delete Image
+            const photoDelete = 'DELETE FROM gallery WHERE uploadedAt = ? AND name = ? AND userID = ?';
             
             //Remove Files Based on File Name
 
             for (let i = 0; i < namesArray.length; i++) {
 
               //Sets Up Query, The keysArray index and namesArray index correspond to the same object (req.body)
-              var deleteQuery = mysql.format(photoDelete, [keysArray[i], namesArray[i]])
-      
+              var deleteQuery = mysql.format(photoDelete, [keysArray[i], namesArray[i], userID])
+
               //Activates Query
               connection.query (deleteQuery, (err, result) => {
-                console.log(err)
-                if (err) reject(err) 
-              }) 
-              //The arrays act as a double check in case the same photo is uploaded multiple times 
-              fs.rmSync(`./public/images/galleryImages/${keysArray[i]}/${namesArray[i]}`, { recursive: true, force: true })
-              try {
-                fs.rmdirSync(`./public/images/galleryImages/${keysArray[i]}`)
-                console.log(`Directory ${keysArray[i]} Deleted`)
-              } catch {
-                console.log(`Directory ${keysArray[i]} Still Contains Photos`)
-              }
+                console.log(err);
+                if (err) reject(err) ;
 
-            }
-            
-          return resolve(connection.release())
-            
+                //DeleteKey Function from galleryDelete.js
+                deleteKey(keysArray[i], namesArray[i], () => {
+                  if (i === namesArray.length - 1) {
+                    connection.release();
+                    res.redirect('/gallery');
+                  }
+                });
+
+                return resolve(result);
+
+              }) 
+              
+          }
+
       })
   
     })
-  
-    res.redirect('/gallery')
-      
+        
     } catch {
-    res.redirect('/editGallery')
-    console.log('ERROR')
+
+      res.redirect('/editGallery');
+      console.log('ERROR');
+
     } 
   
   })
